+++ pollen_vision/pollen_vision/camera_wrappers/depthai/utils.py
@@ -129,3 +129,31 @@ def get_config_file_path(name: str) -> Any:
         if file.stem == name:
             return file.resolve()
     return None
+
+
+def colorizeDepth(frameDepth):
+    invalidMask = frameDepth == 0
+    # Log the depth, minDepth and maxDepth
+    try:
+        minDepth = np.percentile(frameDepth[frameDepth != 0], 3)
+        maxDepth = np.percentile(frameDepth[frameDepth != 0], 95)
+        logDepth = np.log(frameDepth, where=frameDepth != 0)
+        logMinDepth = np.log(minDepth)
+        logMaxDepth = np.log(maxDepth)
+        np.nan_to_num(logDepth, copy=False, nan=logMinDepth)
+        # Clip the values to be in the 0-255 range
+        logDepth = np.clip(logDepth, logMinDepth, logMaxDepth)
+
+        # Interpolate only valid logDepth values, setting the rest based on the mask
+        depthFrameColor = np.interp(logDepth, (logMinDepth, logMaxDepth), (0, 255))
+        depthFrameColor = np.nan_to_num(depthFrameColor)
+        depthFrameColor = depthFrameColor.astype(np.uint8)
+        depthFrameColor = cv2.applyColorMap(depthFrameColor, cv2.COLORMAP_JET)
+        # Set invalid depth pixels to black
+        depthFrameColor[invalidMask] = 0
+    except IndexError:
+        # Frame is likely empty
+        depthFrameColor = np.zeros((frameDepth.shape[0], frameDepth.shape[1], 3), dtype=np.uint8)
+    except Exception as e:
+        raise e
+    return depthFrameColor
